+++ src/timeline/Timeline.cpp
@@ -212,6 +212,10 @@ Timeline::Timeline(const QString &roomId, QObject *parent):
     _permissions(_roomId)
     {
     nhlog::dev()->debug("Timeline created for: \"" + roomId.toStdString() + "\"");
+    connect(cache::client(), &Cache::newReadReceipts,[&](const QString &room_id, const std::vector<QString> &event_ids){
+        if(room_id == _roomId)
+            emit newReadReceipts(event_ids);
+    });
     connect(this,
             &Timeline::newMessageToSend,
             this,
@@ -841,4 +845,8 @@ QString Timeline::viewRawMessage(const QString &id){
     if (!e)
         return "";
     return QString::fromStdString(mtx::accessors::serialize_event(*e).dump(4));
+}
+
+Timeline::UserReceipts Timeline::readReceipts(const QString &event_id){
+    return cache::readReceipts(event_id, _roomId);
 }